requirements.txt
aiogram==3.13.0
aiohttp>=3.8.0
aiofiles>=0.8.0
python-dotenv>=0.19.0
Flask>=2.0.0